+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+### Added
+- Ability to switch on/off adding Playwright project names to code reference via config property `includePlaywrightProjectNameToCodeReference`.
 
 ## [5.0.5] - 2022-07-12
 ### Fixed